More pr feedback
fix a possible bug??
