--- lute/runtime/src/UVStream.cpp
@@ -1,8 +1,0 @@
-#include "lute/UVStream.h"
-
-#include "uv.h"
-
-namespace uvutils
-{
-
-} // namespace uvutils
